refactor(layout): split verticalText.ts into vertical/ module family
One day of vertical-text feature work grew verticalText.ts to 1,218
lines holding six concerns. Split it along its real seams into
vertical/displayTitle.ts (tall/display path + tatechuyoko),
vertical/bodyRuns.ts (body runs, columns, short corroborated
columns), vertical/ruby.ts (classification, association,
attachment), vertical/shared.ts (common types/geometry/constants)
and an index barrel, leaving verticalText.ts as a one-line re-export
shim so no call site churns. Vertical tests moved out of the
3,127-line layout.unit.test.ts into tests/core/vertical/ mirroring
the module split, assertions untouched.

Strictly structural: no logic, threshold, name, or message changes.
Full-output (-f json --layout --geometry) byte identity verified
against a clean pre-refactor build across the vertical corpus
(kokoro3, rashomon, Taishukan, DNC exam, Shugiin, geppo, kyushu) and
horizontal controls (twada, NIST); test count unchanged at 1106.

Implemented by Codex CLI from a written spec; reviewed, independently
verified and integrated by Claude.
