+++ src/core/layout/vertical/index.ts
@@ -1,0 +1,20 @@
+export {
+  compareLayoutBlocks,
+  extractTallVerticalRuns,
+  extractVerticalCjkBlocks,
+} from './displayTitle.js';
+export {
+  extractBodyVerticalCjkRunAnalysis,
+  extractBodyVerticalCjkRunBlocks,
+  rubyAssociationText,
+  verticalRunTextWithRuby,
+} from './ruby.js';
+export {
+  type BodyVerticalCjkRunAnalysis,
+  hasVerticalTextShape,
+  type TallVerticalRun,
+  type VerticalCjkRun,
+  type VerticalCjkRunBlock,
+  type VerticalRubyAssociation,
+  type VerticalRubyBaseRange,
+} from './shared.js';
