+++ tests/core/layout.unit.test.ts
@@ -1,6 +1,5 @@
 import { describe, expect, it } from 'vitest';
 import { buildLayout, markRepeatedBlocks } from '../../src/core/layout/index.js';
-import { extractBodyVerticalCjkRunAnalysis } from '../../src/core/layout/verticalText.js';
 import { detectPageWarnings } from '../../src/core/warnings/index.js';
 import type { LayoutBlock, PageResult, TextSpan } from '../../src/types/index.js';
 
@@ -18,37 +17,6 @@ function span(text: string, x: number, y: number, fontSize: number, width = text
     height: fontSize,
     fontSize,
   };
-}
-
-function verticalGlyphs(text: string, x: number, y: number, fontSize: number, step = fontSize): TextSpan[] {
-  return Array.from(text).map((glyph, index) => span(glyph, x, y + index * step, fontSize, fontSize));
-}
-
-function verticalSpan(
-  text: string,
-  x: number,
-  y: number,
-  fontSize: number,
-  height = Array.from(text).length * fontSize,
-): TextSpan {
-  return {
-    text,
-    x,
-    y,
-    width: fontSize,
-    height,
-    fontSize,
-  };
-}
-
-function verticalGlyphTexts(
-  texts: readonly string[],
-  x: number,
-  y: number,
-  fontSize: number,
-  step = fontSize,
-): TextSpan[] {
-  return texts.map((text, index) => span(text, x, y + index * step, fontSize, fontSize));
 }
 
 describe('buildLayout — heading classification', () => {
@@ -1532,332 +1500,6 @@ describe('buildLayout — multi-column reading order', () => {
     expect(layout.blocks.flatMap((block) => block.lines.map((line) => line.text))).toEqual(['北海', '道東']);
   });
 
-  it('keeps Japanese vertical glyph stacks as separate top-to-bottom blocks', () => {
-    // Japanese slide-title-shaped input from a public government PDF:
-    // pdf.js emits one square-ish glyph per span. A y-row-only layout pass
-    // used to merge the two vertical columns row-wise (`縦 書\n書 籍...`).
-    const spans: TextSpan[] = [
-      span('ネットの横書き', 290, 430, 32, 210),
-      span('縦', 36, 194, 76, 72),
-      span('書', 36, 299, 76, 72),
-      span('き', 36, 405, 76, 72),
-      span('書', 182, 97, 76, 72),
-      span('籍', 182, 202, 76, 72),
-      span('の', 182, 308, 76, 72),
-      span('と', 589, 137, 92, 86),
-    ];
-    const layout = buildLayout(spans, 720);
-    const verticalBlocks = layout.blocks.filter((block) => block.writingMode === 'vertical');
-
-    expect(verticalBlocks.map((block) => block.text)).toEqual(expect.arrayContaining(['縦書き', '書籍の']));
-    expect(verticalBlocks.every((block) => block.lines[0]?.writingMode === 'vertical')).toBe(true);
-    expect(layout.blocks.map((block) => block.text)).not.toContain('縦 書\n書 籍\nき の');
-  });
-
-  it('groups body-sized Japanese vertical glyph runs right-to-left', () => {
-    const rightColumn = Array.from('質問主意書').map((text, index) => span(text, 300, 100 + index * 8, 8, 8));
-    const leftColumn = Array.from('国会質疑中').map((text, index) => span(text, 276, 100 + index * 8, 8, 8));
-
-    const layout = buildLayout([...leftColumn, ...rightColumn], 595);
-    const verticalBlocks = layout.blocks.filter((block) => block.writingMode === 'vertical');
-
-    expect(verticalBlocks).toHaveLength(1);
-    expect(verticalBlocks[0].text).toBe('質問主意書\n国会質疑中');
-    expect(verticalBlocks[0].lines.map((line) => line.text)).toEqual(['質問主意書', '国会質疑中']);
-    expect(verticalBlocks[0].lines.every((line) => line.writingMode === 'vertical')).toBe(true);
-  });
-
-  it('accepts short vertical columns with normalized ellipsis leaders when body columns corroborate them', () => {
-    const rightColumn = verticalGlyphs('右側本文甲乙丙丁', 300, 100, 12);
-    const shortColumn = verticalGlyphTexts(['そ', 'れ', 'と', 'も', '...', '...'], 279, 124, 12);
-    const leftColumn = verticalGlyphs('左側本文甲乙丙丁', 258, 100, 12);
-    const spans = [...leftColumn, ...shortColumn, ...rightColumn];
-
-    const analysis = extractBodyVerticalCjkRunAnalysis(spans);
-    expect(analysis.blocks).toHaveLength(1);
-    expect(analysis.blocks[0].columns.map((column) => column.spans.map((item) => item.text).join(''))).toEqual([
-      '右側本文甲乙丙丁',
-      'それとも......',
-      '左側本文甲乙丙丁',
-    ]);
-
-    const layout = buildLayout(spans, 595);
-    const verticalBlocks = layout.blocks.filter((block) => block.writingMode === 'vertical');
-    expect(verticalBlocks).toHaveLength(1);
-    expect(verticalBlocks[0].lines.map((line) => line.text)).toEqual([
-      '右側本文甲乙丙丁',
-      'それとも......',
-      '左側本文甲乙丙丁',
-    ]);
-    expect(verticalBlocks[0].lines.every((line) => line.writingMode === 'vertical')).toBe(true);
-  });
-
-  it('does not accept the same short ellipsis column without vertical body context', () => {
-    const shortColumn = verticalGlyphTexts(['そ', 'れ', 'と', 'も', '...', '...'], 120, 100, 12);
-
-    const analysis = extractBodyVerticalCjkRunAnalysis(shortColumn);
-    expect(analysis.blocks).toEqual([]);
-
-    const layout = buildLayout(shortColumn, 300);
-    expect(layout.blocks.some((block) => block.writingMode === 'vertical')).toBe(false);
-  });
-
-  it('classifies adjacent half-size ruby stacks without adding them to body vertical blocks', () => {
-    const rightBody = verticalGlyphs('私東京で相当の地位を得たい', 300, 100, 12);
-    const leftBody = verticalGlyphs('から宜しく頼む', 279, 100, 12);
-    const ruby = verticalGlyphs('わたくし', 312, 94, 6);
-    const spans = [...rightBody, ...ruby, ...leftBody];
-
-    const analysis = extractBodyVerticalCjkRunAnalysis(spans);
-    expect(analysis.rubySpans.map((item) => item.text).join('')).toBe('わたくし');
-    expect(analysis.rubyAssociations).toHaveLength(1);
-    expect(analysis.rubyAssociations[0].baseSpans.map((item) => item.text).join('')).toBe('私');
-    expect(analysis.blocks).toHaveLength(1);
-    expect(analysis.blocks[0].columns.map((column) => column.spans.map((item) => item.text).join(''))).toEqual([
-      '私東京で相当の地位を得たい',
-      'から宜しく頼む',
-    ]);
-
-    const layout = buildLayout(spans, 595);
-    const verticalBlocks = layout.blocks.filter((block) => block.writingMode === 'vertical');
-    expect(verticalBlocks.map((block) => block.text)).toContain(
-      '私《わたくし》東京で相当の地位を得たい\nから宜しく頼む',
-    );
-  });
-
-  it('does not readmit half-size short ruby stacks through contextual short-run detection', () => {
-    const rightBody = verticalGlyphs('東京で相当の地位を得たい', 300, 100, 12);
-    const leftBody = verticalGlyphs('から宜しく頼む', 279, 100, 12);
-    const ruby = verticalGlyphs('ふりがな', 312, 124, 6);
-    const spans = [...rightBody, ...ruby, ...leftBody];
-
-    const analysis = extractBodyVerticalCjkRunAnalysis(spans);
-    expect(analysis.rubySpans.map((item) => item.text).join('')).toBe('ふりがな');
-    expect(analysis.blocks.flatMap((block) => block.columns).flatMap((column) => column.spans)).not.toEqual(
-      expect.arrayContaining(ruby),
-    );
-  });
-
-  it('keeps uniformly small vertical columns as body when no larger adjacent column exists', () => {
-    const rightBody = verticalGlyphs('質問主意書', 120, 100, 6);
-    const leftBody = verticalGlyphs('国会質疑中', 108, 100, 6);
-    const spans = [...rightBody, ...leftBody];
-
-    const analysis = extractBodyVerticalCjkRunAnalysis(spans);
-    expect(analysis.rubySpans).toEqual([]);
-    expect(analysis.blocks[0].columns.map((column) => column.spans.map((item) => item.text).join(''))).toEqual([
-      '質問主意書',
-      '国会質疑中',
-    ]);
-
-    const layout = buildLayout(spans, 300);
-    expect(layout.blocks.filter((block) => block.writingMode === 'vertical').map((block) => block.text)).toEqual([
-      '質問主意書\n国会質疑中',
-    ]);
-  });
-
-  it('does not classify a small vertical stack far from a larger body column as ruby', () => {
-    const body = verticalGlyphs('東京で相当の地位を得たい', 300, 100, 12);
-    const farSmall = verticalGlyphs('かな', 330, 124, 6);
-
-    const analysis = extractBodyVerticalCjkRunAnalysis([...body, ...farSmall]);
-
-    expect(analysis.rubySpans).toEqual([]);
-    expect(analysis.blocks[0].columns.map((column) => column.spans.map((item) => item.text).join(''))).toEqual([
-      '東京で相当の地位を得たい',
-    ]);
-  });
-
-  it('classifies a two-glyph ruby stack when it is adjacent to a larger body column', () => {
-    const body = verticalGlyphs('東京で相当の地位を得たい', 300, 100, 12);
-    const ruby = verticalGlyphs('わた', 312, 124, 6);
-
-    const analysis = extractBodyVerticalCjkRunAnalysis([...body, ...ruby]);
-
-    expect(analysis.rubySpans.map((item) => item.text).join('')).toBe('わた');
-    expect(analysis.blocks[0].columns.map((column) => column.spans.map((item) => item.text).join(''))).toEqual([
-      '東京で相当の地位を得たい',
-    ]);
-  });
-
-  it('attaches ruby to an internal range of a multi-character vertical body span', () => {
-    const body = verticalSpan('甲朱雀乙丙丁戊己庚辛', 300, 100, 12);
-    const ruby = verticalSpan('すざく', 312, 112, 6, 24);
-    const spans = [body, ruby];
-
-    const analysis = extractBodyVerticalCjkRunAnalysis(spans);
-    expect(analysis.rubySpans).toEqual([ruby]);
-    expect(analysis.rubyAssociations).toHaveLength(1);
-    expect(analysis.rubyAssociations[0].baseRanges).toMatchObject([{ span: body, start: 1, end: 3 }]);
-
-    const layout = buildLayout(spans, 595);
-    expect(layout.blocks.filter((block) => block.writingMode === 'vertical').map((block) => block.text)).toEqual([
-      '甲朱雀《すざく》乙丙丁戊己庚辛',
-    ]);
-  });
-
-  it('does not treat x-aligned multi-character CJK spans as body vertical writing', () => {
-    const spans: TextSpan[] = ['質問', '主意', '書面', '回答', '資料'].map((text, index) =>
-      span(text, 80, 100 + index * 8, 8, 16),
-    );
-
-    const layout = buildLayout(spans, 300);
-
-    expect(layout.blocks.some((block) => block.writingMode === 'vertical')).toBe(false);
-  });
-
-  it('does not treat short x-aligned CJK glyph runs as body vertical writing', () => {
-    const spans: TextSpan[] = Array.from('質問書').map((text, index) => span(text, 80, 100 + index * 8, 8, 8));
-
-    const layout = buildLayout(spans, 300);
-
-    expect(layout.blocks.some((block) => block.writingMode === 'vertical')).toBe(false);
-  });
-
-  it('does not treat short x-aligned table labels near horizontal spans as body vertical writing', () => {
-    const label = verticalGlyphs('注記表', 80, 100, 8, 8);
-    const spans: TextSpan[] = [span('横書きの表見出し', 48, 92, 8, 80), span('別の表セル', 96, 108, 8, 48), ...label];
-
-    const layout = buildLayout(spans, 300);
-
-    expect(layout.blocks.some((block) => block.writingMode === 'vertical')).toBe(false);
-  });
-
-  it('marks tall CJK spans as vertical columns in right-to-left order', () => {
-    // PDF.js vertical.pdf-shaped input: each vertical column arrives as
-    // one tall span whose text is already top-to-bottom.
-    const spans: TextSpan[] = [
-      span('あいうえお', 233.86, 21.97, 9.21, 9.21),
-      span('日本語', 218.27, 21.97, 9.21, 9.21),
-    ];
-    spans[0].height = 46.06;
-    spans[1].height = 27.64;
-
-    const layout = buildLayout(spans, 300);
-    expect(layout.blocks.map((block) => block.text)).toEqual(['あいうえお', '日本語']);
-    expect(layout.blocks.every((block) => block.writingMode === 'vertical')).toBe(true);
-    expect(layout.blocks.every((block) => block.lines[0]?.writingMode === 'vertical')).toBe(true);
-  });
-
-  it('marks tall vertical spans with leading ASCII parens when they contain CJK text', () => {
-    const spans: TextSpan[] = [span('(1935)年5月、新聞にこんな', 467.1, 615.32, 9.5, 9.5)];
-    spans[0].height = 152;
-
-    const layout = buildLayout(spans, 595);
-
-    expect(layout.blocks).toHaveLength(1);
-    expect(layout.blocks[0].text).toBe('(1935)年5月、新聞にこんな');
-    expect(layout.blocks[0].writingMode).toBe('vertical');
-    expect(layout.blocks[0].lines[0]?.writingMode).toBe('vertical');
-  });
-
-  it('stitches tatechuyoko digit fragments into the same vertical layout line', () => {
-    const spans: TextSpan[] = [
-      span('昭和', 100, 100, 10, 10),
-      span('10', 95, 119, 10, 10),
-      span('(1935)年5月、新聞にこんな', 100, 130, 10, 10),
-    ];
-    spans[0].height = 20;
-    spans[2].height = 160;
-
-    const layout = buildLayout(spans, 300);
-    const verticalBlocks = layout.blocks.filter((block) => block.writingMode === 'vertical');
-
-    expect(verticalBlocks).toHaveLength(1);
-    expect(verticalBlocks[0].text).toBe('昭和10(1935)年5月、新聞にこんな');
-    expect(verticalBlocks[0].lines).toHaveLength(1);
-    expect(verticalBlocks[0].lines[0]).toMatchObject({
-      text: '昭和10(1935)年5月、新聞にこんな',
-      writingMode: 'vertical',
-    });
-  });
-
-  it('does not classify horizontal digit rows as vertical writing', () => {
-    const spans: TextSpan[] = [span('10', 50, 100, 10, 10), span('20', 70, 100, 10, 10), span('30', 90, 100, 10, 10)];
-
-    const layout = buildLayout(spans, 300);
-
-    expect(layout.blocks.some((block) => block.writingMode === 'vertical')).toBe(false);
-    expect(layout.blocks[0].lines[0]?.text).toBe('10 20 30');
-  });
-
-  it('does not classify tall ASCII-only spans as vertical CJK writing', () => {
-    const spans: TextSpan[] = [span('https://example.com/path', 100, 100, 10, 10)];
-    spans[0].height = 180;
-
-    const layout = buildLayout(spans, 300);
-
-    expect(layout.blocks.some((block) => block.writingMode === 'vertical')).toBe(false);
-  });
-
-  it('marks page-edge vertical navigation tabs as repeated chrome', () => {
-    const spans: TextSpan[] = [
-      span('本文の見出し', 65, 54, 18, 120),
-      span('本文の一行目がここにあります', 62, 84, 10, 220),
-      span('第', 22, 126, 8.5, 8.5),
-      span('1', 15, 132, 19.8, 13.1),
-      span('章', 22, 155, 8.5, 8.5),
-      span('働き方改革の推進などを通じた労働環境の整備など', 23, 181, 9.2, 9.2),
-      span('本文の二行目がここにあります', 62, 140, 10, 220),
-    ];
-    spans[5].height = 212;
-
-    const layout = buildLayout(spans, 595.28, 841.89);
-    const chromeTexts = layout.blocks.filter((block) => block.repeated).map((block) => block.text);
-
-    expect(chromeTexts).toEqual(expect.arrayContaining(['第', '1', '章']));
-    expect(chromeTexts).toContain('働き方改革の推進などを通じた労働環境の整備など');
-    expect(layout.blocks.find((block) => block.text === '本文の一行目がここにあります')?.repeated).toBeUndefined();
-  });
-
-  it('does not treat aligned first glyphs of horizontal CJK lines as vertical writing', () => {
-    const spans: TextSpan[] = [
-      span('日', 50, 50, 12, 12),
-      span('本', 63, 50, 12, 12),
-      span('語', 76, 50, 12, 12),
-      span('日', 50, 68, 12, 12),
-      span('本', 63, 68, 12, 12),
-      span('語', 76, 68, 12, 12),
-      span('日', 50, 86, 12, 12),
-      span('本', 63, 86, 12, 12),
-      span('語', 76, 86, 12, 12),
-    ];
-    const layout = buildLayout(spans);
-
-    expect(layout.blocks.some((block) => block.writingMode === 'vertical')).toBe(false);
-    expect(layout.blocks.flatMap((block) => block.lines.map((line) => line.text))).toEqual([
-      '日本語',
-      '日本語',
-      '日本語',
-    ]);
-  });
-
-  it('does not extract small horizontal CJK labels with wide spacing as vertical blocks', () => {
-    // Table/list-shaped Japanese text can repeat short labels at the same
-    // x across rows while using a deliberate full-width-ish gap inside
-    // each row. Those rows should stay horizontal, not get stripped into a
-    // top-to-bottom label.
-    const spans: TextSpan[] = [
-      span('序', 50, 50, 12, 12),
-      span('文', 66.64, 50, 12, 12), // gap ≈ 0.72 × fontSize
-      span('本', 90, 50, 12, 40),
-      span('序', 50, 68, 12, 12),
-      span('文', 66.64, 68, 12, 12),
-      span('本', 90, 68, 12, 40),
-      span('序', 50, 86, 12, 12),
-      span('文', 66.64, 86, 12, 12),
-      span('本', 90, 86, 12, 40),
-    ];
-    const layout = buildLayout(spans);
-
-    expect(layout.blocks.some((block) => block.writingMode === 'vertical')).toBe(false);
-    expect(layout.blocks.flatMap((block) => block.lines.map((line) => line.text))).toEqual([
-      '序 文 本',
-      '序 文 本',
-      '序 文 本',
-    ]);
-  });
-
   it('keeps a semantic space before a URL when the visual gap is narrowly below the default threshold', () => {
     // ACL-style font-run boundary: the gap before the URL is just below
     // 0.25x fontSize, but the token is visually and semantically
@@ -3119,21 +2761,6 @@ describe('buildLayout — multi-column reading order', () => {
 
     expect(lines).toEqual(['Natural Language', 'Processing']);
     expect(layout.blocks[0].text).toBe('Natural Language\nProcessing');
-  });
-
-  it('does not merge vertical side labels into horizontal text lines', () => {
-    const spans: TextSpan[] = [
-      { text: '(Version 2)', x: 114, y: 200, width: 45, height: 10, fontSize: 10 },
-      { text: 'arXiv:2106.09685v2 [cs.CL] 16 Oct 2021', x: 12, y: 214, width: 20, height: 346, fontSize: 20 },
-      span('Abstract body text starts in the main column.', 144, 265, 10, 220),
-    ];
-    const layout = buildLayout(spans, 612);
-    const sidebar = layout.blocks.find((b) => b.text.includes('arXiv'));
-    const version = layout.blocks.find((b) => b.text.includes('Version'));
-
-    expect(sidebar?.text).toBe('arXiv:2106.09685v2 [cs.CL] 16 Oct 2021');
-    expect(sidebar?.width).toBe(20);
-    expect(version?.text).toBe('(Version 2)');
   });
 
   it('does not falsely detect columns when only one block sits at a different x', () => {
